ntp: fix clang warning

+++ ntp_core.c
@@ -1521,8 +1521,8 @@ receive_packet(NCR_Instance inst, NTP_Local_Address *local_addr,
     inst->valid_rx = 1;
   }
 
-  if (sample_rx_tss < 0 || sample_rx_tss >= sizeof (tss_chars) ||
-      inst->local_tx.source < 0 || inst->local_tx.source >= sizeof (tss_chars))
+  if ((int)sample_rx_tss < 0 || sample_rx_tss >= sizeof (tss_chars) ||
+      (int)inst->local_tx.source < 0 || inst->local_tx.source >= sizeof (tss_chars))
     assert(0);
 
   DEBUG_LOG(LOGF_NtpCore, "NTP packet lvm=%o stratum=%d poll=%d prec=%d root_delay=%f root_disp=%f refid=%"PRIx32" [%s]",
